+++ src/interfaces/API/IPokemonMovie.ts
@@ -1,10 +1,3 @@
-export interface IPokemonMoviesResponse {
-  page: number;
-  results: IPokemonMovie[];
-  total_pages: number;
-  total_results: number;
-}
-
 export interface IPokemonMovie {
   backdrop_path: string;
   id: number;
