+++ src/interfaces/API/IPokemonMoviesResponse.ts
@@ -1,0 +1,8 @@
+import { IPokemonMovie } from "./IPokemonMovie";
+
+export interface IPokemonMoviesResponse {
+  page: number;
+  results: IPokemonMovie[];
+  total_pages: number;
+  total_results: number;
+}
