+++ src/interfaces/API/index.ts
@@ -1,4 +1,5 @@
 export type { IApiResponse } from './IApiResponse';
 export type { IPokemon } from './IPokemon';
 export type { IPokemonDetail } from './IPokemonDetail';
-export type { IPokemonMoviesResponse, IPokemonMovie } from './IPokemonMovies';
+export type { IPokemonMovie } from './IPokemonMovie';
+export type { IPokemonMoviesResponse } from './IPokemonMoviesResponse';
